feat: moved sinpedata and kilometer price to ConfigItem
